requirements.txt
pandas~=2.1.4
datasets~=2.18.0
requests~=2.31.0
tqdm~=4.66.1
openpyxl
tensorflow~=2.16.0rc0
beautifulsoup4
nltk~=3.8.1
pyspellchecker
